n8n_mcp_server/requirements.txt
fastmcp>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0

